# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample FHIR Patient API server in Go using Gin.",
        "title": "Go FHIR Demo API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/external-patients": {
            "get": {
                "description": "Searches for patient resources on an external FHIR server based on query parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ExternalPatients"
                ],
                "summary": "Search for external patients",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FHIR search parameters (e.g., name=John,birthdate=1990-01-01)",
                        "name": "_query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved search results",
                        "schema": {
                            "$ref": "#/definitions/fhir.Bundle"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new patient resource on an external FHIR server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ExternalPatients"
                ],
                "summary": "Create an external patient",
                "parameters": [
                    {
                        "description": "Patient resource to create (FHIR-compliant JSON)",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created patient",
                        "schema": {
                            "$ref": "#/definitions/fhir.Patient"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/external-patients/{id}": {
            "get": {
                "description": "Retrieves a patient resource from an external FHIR server by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ExternalPatients"
                ],
                "summary": "Get an external patient by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved patient",
                        "schema": {
                            "$ref": "#/definitions/fhir.Patient"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/external-patients/{id}/cached": {
            "get": {
                "description": "Retrieves a patient resource from an external FHIR server by its ID with Redis caching",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ExternalPatients"
                ],
                "summary": "Get an external patient by ID with caching",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved patient",
                        "schema": {
                            "$ref": "#/definitions/fhir.Patient"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/external-patients/{id}/delayed": {
            "get": {
                "description": "Retrieves a patient resource from an external FHIR server by its ID with configurable timeout",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ExternalPatients"
                ],
                "summary": "Get an external patient by ID with timeout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Timeout in seconds (default: 10)",
                        "name": "timeout",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved patient",
                        "schema": {
                            "$ref": "#/definitions/fhir.Patient"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fhir.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "line": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "period": {
                    "$ref": "#/definitions/fhir.Period"
                },
                "postalCode": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/fhir.AddressType"
                },
                "use": {
                    "$ref": "#/definitions/fhir.AddressUse"
                }
            }
        },
        "fhir.AddressType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "AddressTypePostal",
                "AddressTypePhysical",
                "AddressTypeBoth"
            ]
        },
...
        "fhir.AdministrativeGender": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "AdministrativeGenderMale",
                "AdministrativeGenderFemale",
                "AdministrativeGenderOther",
                "AdministrativeGenderUnknown"
            ]
        },
...
        "fhir.Bundle": {
            "type": "object",
            "properties": {
                "entry": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.BundleEntry"
                    }
                },
                "id": {
                    "type": "string"
                },
                "identifier": {
                    "$ref": "#/definitions/fhir.Identifier"
                },
                "implicitRules": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "link": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.BundleLink"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/fhir.Meta"
                },
                "signature": {
                    "$ref": "#/definitions/fhir.Signature"
                },
                "timestamp": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/fhir.BundleType"
                }
            }
        },
        "fhir.BundleEntry": {
            "type": "object",
            "properties": {
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "fullUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.BundleLink"
                    }
                },
                "modifierExtension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "request": {
                    "$ref": "#/definitions/fhir.BundleEntryRequest"
                },
                "resource": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "response": {
                    "$ref": "#/definitions/fhir.BundleEntryResponse"
                },
                "search": {
                    "$ref": "#/definitions/fhir.BundleEntrySearch"
                }
            }
        },
...
        "fhir.BundleLink": {
            "type": "object",
            "properties": {
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "modifierExtension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "relation": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "fhir.BundleType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8
            ],
            "x-enum-varnames": [
                "BundleTypeDocument",
                "BundleTypeMessage",
                "BundleTypeTransaction",
                "BundleTypeTransactionResponse",
                "BundleTypeBatch",
                "BundleTypeBatchResponse",
                "BundleTypeHistory",
                "BundleTypeSearchset",
                "BundleTypeCollection"
            ]
        },
...
        "fhir.Identifier": {
            "type": "object",
            "properties": {
                "assigner": {
                    "$ref": "#/definitions/fhir.Reference"
                },
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "period": {
                    "$ref": "#/definitions/fhir.Period"
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/fhir.CodeableConcept"
                },
                "use": {
                    "$ref": "#/definitions/fhir.IdentifierUse"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "fhir.Meta": {
            "type": "object",
            "properties": {
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "lastUpdated": {
                    "type": "string"
                },
                "profile": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "security": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Coding"
                    }
                },
                "source": {
                    "type": "string"
                },
                "tag": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Coding"
                    }
                },
                "versionId": {
                    "type": "string"
                }
            }
        },
...
        "fhir.Patient": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "address": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Address"
                    }
                },
                "birthDate": {
                    "type": "string"
                },
                "communication": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.PatientCommunication"
                    }
                },
                "contact": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.PatientContact"
                    }
                },
                "deceasedBoolean": {
                    "type": "boolean"
                },
                "deceasedDateTime": {
                    "type": "string"
                },
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "gender": {
                    "$ref": "#/definitions/fhir.AdministrativeGender"
                },
                "generalPractitioner": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Reference"
                    }
                },
                "id": {
                    "type": "string"
                },
                "identifier": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Identifier"
                    }
                },
                "implicitRules": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "link": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.PatientLink"
                    }
                },
                "managingOrganization": {
                    "$ref": "#/definitions/fhir.Reference"
                },
                "maritalStatus": {
                    "$ref": "#/definitions/fhir.CodeableConcept"
                },
                "meta": {
                    "$ref": "#/definitions/fhir.Meta"
                },
                "modifierExtension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "multipleBirthBoolean": {
                    "type": "boolean"
                },
                "multipleBirthInteger": {
                    "type": "integer"
                },
                "name": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.HumanName"
                    }
                },
                "photo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Attachment"
                    }
                },
                "telecom": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.ContactPoint"
                    }
                },
                "text": {
                    "$ref": "#/definitions/fhir.Narrative"
                }
            }
        },
        "fhir.PatientCommunication": {
            "type": "object",
            "properties": {
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "$ref": "#/definitions/fhir.CodeableConcept"
                },
                "modifierExtension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "preferred": {
                    "type": "boolean"
                }
            }
        },
        "fhir.PatientContact": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/fhir.Address"
                },
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "gender": {
                    "$ref": "#/definitions/fhir.AdministrativeGender"
                },
                "id": {
                    "type": "string"
                },
                "modifierExtension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "name": {
                    "$ref": "#/definitions/fhir.HumanName"
                },
                "organization": {
                    "$ref": "#/definitions/fhir.Reference"
                },
                "period": {
                    "$ref": "#/definitions/fhir.Period"
                },
                "relationship": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.CodeableConcept"
                    }
                },
                "telecom": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.ContactPoint"
                    }
                }
            }
        },
        "fhir.PatientLink": {
            "type": "object",
            "properties": {
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "modifierExtension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "other": {
                    "$ref": "#/definitions/fhir.Reference"
                },
                "type": {
                    "$ref": "#/definitions/fhir.LinkType"
                }
            }
        },
...
        "fhir.Reference": {
            "type": "object",
            "properties": {
                "display": {
                    "type": "string"
                },
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "identifier": {
                    "$ref": "#/definitions/fhir.Identifier"
                },
                "reference": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "fhir.Signature": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "extension": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Extension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "onBehalfOf": {
                    "$ref": "#/definitions/fhir.Reference"
                },
                "sigFormat": {
                    "type": "string"
                },
                "targetFormat": {
                    "type": "string"
                },
                "type": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fhir.Coding"
                    }
                },
                "when": {
                    "type": "string"
                },
                "who": {
                    "$ref": "#/definitions/fhir.Reference"
                }
            }
        },
...
    }
}